requirements.txt
crewai[tools]>=0.70.0
crewai-tools>=0.8.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
python-telegram-bot>=20.5
langchain-community>=0.0.10
langchain-google-community>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
getpass3>=1.0.0